Separate functions
Add quadratic functions to the quadratic.cpp
